research(e5): make the staged workflow executable rather than documented
E5's governance was Markdown plus a manifest-chain checker. Markdown does not stop
anybody editing a protocol after the results are known - which is precisely how E4
ended up, with protocol, implementation, results and post-hoc audit entering public
history in one commit. This replaces the promise with mechanics.

research/e5/harness/e5_harness.py defines nine stages (0 protocol .. 8 final-report),
a manifest schema, frozen writes, and a verifier. Each manifest records sequence,
previous manifest hash, UTC timestamp, git commit and worktree cleanliness,
environment (interpreter, platform, requirements.lock hash, required and observed
environment variables), the hashes of the frozen protocol, config and representations
documents, the frozen model identity hash, artifact hashes with roles and byte counts,
and stage counts.

The verifier enforces:
- stages form a complete prefix, in order, with non-decreasing timestamps;
- every artifact still hashes to the recorded value;
- the manifest chain: stage N's recorded previous hash equals the canonical hash of
  stage N-1;
- the protocol, config and representations recorded at freeze still hash to the same
  values, so editing any of them after a freeze fails;
- the frozen model identity is unchanged;
- outcome-role artifacts appear only from stage 5;
- any recorded outcome path that was asserted inaccessible and later appears fails;
- cohort disjointness that is not PROVEN fails closed, because previous_270.json is
  still unpublished.

A manifest's own hash is stored inside it, so editing a manifest after it was written
is detected as well. write_manifest refuses to overwrite, so a stage cannot be
silently re-run.

Entry points:
- research/e5/harness/verify_e5.py - one command; missing stages are reported as
  absent, not as failures, so it is safe in CI today.
- research/e5/harness/verify_environment.py - freeze or check the environment
  manifest. This is the direct fix for E4's undocumented
  FINRISK_SEC_MAX_ARCHIVE_MEMBER_BYTES override.

Also adds assert_single_company_context, the mechanical guard for the
one-company-per-semantic-request rule, and assert_outcome_inaccessible for the
prediction-freeze assertion. No stage beyond a test fixture is frozen, and no result
is produced.

Housekeeping:
- CI now runs `ruff check backend tests scripts research`; research modules were made
  clean (unused imports, int(round(...)), zip -> itertools.pairwise, naive ISO parsing).
  The int(round()) -> round() change is value-neutral: recomputing 54 committed
  analytic power rows reproduces power_analysis.json exactly (max difference 0.0).
- pythonpath now exposes research/e5/harness so tests import it normally.

31 new tests covering chain breakage, skipped stages, protocol/config/model-identity
mutation, manifest self-hash tampering, outcome artifacts before unlock, leaked
outcome paths, fail-closed disjointness, environment drift, single-company context
rejection, and both CLI entry points.

+++ research/e5/protocol/power_analysis.py
@@ -34,13 +34,12 @@ import random
 import sys
 from dataclasses import asdict, dataclass
 from pathlib import Path
-from statistics import mean
 
 AUDIT_DIR = Path(__file__).resolve().parents[2] / "e4_statistical_audit"
 sys.path.insert(0, str(AUDIT_DIR))
 
-from e4s_stats import delong_paired, normal_cdf  # noqa: E402
-from method_calibration import generate_paired  # noqa: E402
+from e4s_stats import delong_paired, normal_cdf
+from method_calibration import generate_paired
 
 # Planning inputs. Every one of these is either an E4 *operational* rate (allowed for
 # planning) or an a-priori design choice. None is an E4 test-set optimum.
@@ -111,7 +110,7 @@ def analytic_power(
     ``Var(dAUC) = Var(A1) + Var(A2) - 2 rho sqrt(Var(A1) Var(A2))``. The challenger's
     variance is evaluated at ``auc_reference + delta``.
     """
-    n_events = int(round(n_evaluable * event_prevalence))
+    n_events = round(n_evaluable * event_prevalence)
     n_non_events = n_evaluable - n_events
     if n_events < 2 or n_non_events < 2:
         return None
@@ -149,7 +148,7 @@ def analytic_power_table(
             key = f"delta={delta:.2f}|rho={correlation:.2f}"
             row: dict[str, float] = {}
             for selected in selected_grid:
-                n_evaluable = int(round(selected * verified_rate * schema_success_rate))
+                n_evaluable = round(selected * verified_rate * schema_success_rate)
                 result = analytic_power(n_evaluable, delta, correlation, auc_reference, event_prevalence, alpha)
                 row[f"selected={selected}"] = result["power"] if result else float("nan")
             table[key] = row
@@ -158,8 +157,8 @@ def analytic_power_table(
 
 def simulate_power(inputs: PowerInputs, replicates: int = 300, seed: int = 20260925) -> dict:
     counts = expected_counts(inputs)
-    n_evaluable = max(2, int(round(counts["expected_evaluable"])))
-    n_events = max(1, int(round(n_evaluable * inputs.event_prevalence)))
+    n_evaluable = max(2, round(counts["expected_evaluable"]))
+    n_events = max(1, round(n_evaluable * inputs.event_prevalence))
     n_non_events = max(1, n_evaluable - n_events)
     if n_events + n_non_events < 4:
         return {**asdict(inputs), "replicates": replicates, "power": None, "reason": "cohort too small"}
